move App.js to root
