+++ packages/react-scripts/template/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import Root from '@fs/zion-root'
 import { addTranslations } from '@fs/zion-locale'
 import { Router } from '@fs/zion-router'
-import App from './components/App'
+import App from './App'
 import * as serviceWorker from './serviceWorker'
 import translations from './locales'
